Rename util/string.h to util/strings.h due to potential system conflicts, fixes #356
