+++ apps/makescene/makescene.cc
@@ -31,7 +31,7 @@
 #include "util/system.h"
 #include "util/timer.h"
 #include "util/arguments.h"
-#include "util/string.h"
+#include "util/strings.h"
 #include "util/file_system.h"
 #include "util/tokenizer.h"
 #include "mve/bundle.h"
