+++ apps/meshalign/meshlab_alignment.cc
@@ -13,7 +13,7 @@
 #include <cerrno>
 
 #include "util/exception.h"
-#include "util/string.h"
+#include "util/strings.h"
 #include "util/file_system.h"
 #include "util/tokenizer.h"
 #include "mve/mesh_io.h"
