+++ apps/umve/scene_addins/mesh_list.cc
@@ -19,7 +19,7 @@
 #include <QColorDialog>
 
 #include "util/file_system.h"
-#include "util/string.h"
+#include "util/strings.h"
 #include "mve/mesh.h"
 #include "mve/mesh_io.h"
 #include "mve/mesh_tools.h"
