+++ libs/fssr/iso_octree.cc
@@ -18,7 +18,7 @@
 #include <limits>
 
 #include "util/timer.h"
-#include "util/string.h"
+#include "util/strings.h"
 #include "fssr/basis_function.h"
 #include "fssr/sample.h"
 #include "fssr/iso_octree.h"
