+++ libs/mve/mesh_io.cc
@@ -9,7 +9,7 @@
 
 #include <stdexcept>
 
-#include "util/string.h"
+#include "util/strings.h"
 #include "mve/mesh.h"
 #include "mve/mesh_io.h"
 #include "mve/mesh_io_ply.h"
