+++ libs/ogl/check_gl_error.h
@@ -14,7 +14,7 @@
 
 #include "ogl/defines.h"
 #include "ogl/opengl.h"
-#include "util/string.h"
+#include "util/strings.h"
 
 OGL_NAMESPACE_BEGIN
