+++ libs/util/arguments.cc
@@ -13,7 +13,7 @@
 #include <iomanip>
 #include <limits>
 
-#include "util/string.h"
+#include "util/strings.h"
 #include "util/tokenizer.h"
 #include "util/arguments.h"
