+++ libs/util/strings.h
